Rename TransactionListScene source file
